requirements.txt
# Minimal requirements for step-by-step implementation

# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0

# ART framework - Ready for installation
openpipe-art>=0.4.9

# LangGraph integration - Optional (using mock for now)
# langgraph>=0.2.0
# langchain-core>=0.3.0

# Optional for visualization
# plotly>=5.0.0
# jupyter>=1.0.0

# Development
pytest>=6.0.0
black>=22.0.0